+++ NEWS.md
@@ -8,3 +8,8 @@
 
 * added Wendland (1) covariance function
 * added "tappered" Matern covariance function
+
+# smile 1.0.3
+
+* fixed Wendland (1) covariance function
+* support for sparse covariance matrices through the `Matrix` package